Feat: MIC-55/56/57/58/59/61 — full BTC-USD paper trading support
- MIC-55: Remove crypto skip guards; add _alpaca_symbol() helper (BTC→BTC/USD);
  fractional qty for crypto; GTC time-in-force for 24/7 markets;
  normalize Alpaca symbols on all exit/fill paths
- MIC-56: get_snapshots() now returns crypto prices via yfinance fetch_quotes()
  alongside Alpaca equity snapshots
- MIC-57: _scan_loop() adds crypto-only scan every 6h on weekends (0/6/12/18 ET);
  run_daily_scan() accepts optional tickers param
- MIC-58: _polling_loop() and _housekeeping_loop() continue running on weekends
  when crypto positions are open (_has_open_crypto_positions helper)
- MIC-59: section_eod_exits_btc() regression suite with _TEST_BTC_* fake positions;
  validates rv_pctile=0 does not spuriously trigger vol_exhaustion
- MIC-61: test_all.py --ticker BTC runs full E2E path: preflight/lifecycle/intraday/EOD
  with crypto-aware qty check; section_scanner handles BTC ticker

+++ src/trading/engine.py
@@ -104,7 +104,7 @@ class ExitResult(BaseModel):
 class FillNotification(BaseModel):
     ticker: str
     fill_price: float
-    qty: int | None = None
+    qty: float | None = None
     stop_price: float = 0.0
     target_price: float = 0.0
